.openclaw/workspace/projects/Axe-Capital/step4-alpha-hunter/pyproject.toml
[project]
name = "axe-alpha-hunter"
version = "0.1.0"
description = "Axe Capital Step 4: Alpha Hunter proactive scan"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "yfinance>=0.2.50",
  "ddgs>=9.5.0",
  "pandas>=2.2",
  "playwright>=1.52.0",
]

[project.scripts]
axe-alpha-scan = "axe_alpha.cli:main"
